Retructure class to set state in wrapper
